* build - start of the mongoDB connection
