+++ src/routes/index.js
@@ -3,10 +3,19 @@ import userRouter from "./userRouter.js";
 
 const routes = (app) =>
 {
+    app.route("/").get((req, res) => {
+        res.status(200).send("FrutaControl!");
+    });
+
     app.use(
         express.json(),
-        userRouter
+        userRouter,
+        notFound
     );
+};
+
+const notFound = (req, res) => {
+   return res.status(404).send({Message: "Page not found"});
 }
 
 export default routes
